mod: removed functions from .hpp files to .cpp
